Revert mypy changes (#1836)
* Revert mypy changes

* Remove unused import

+++ torchgeo/datasets/potsdam.py
@@ -192,8 +192,7 @@ class Potsdam2D(NonGeoDataset):
         path = self.files[index]["image"]
         with rasterio.open(path) as f:
             array = f.read()
-            tensor = torch.from_numpy(array)
-            tensor = tensor.float()
+            tensor = torch.from_numpy(array).float()
             return tensor
 
     def _load_target(self, index: int) -> Tensor:
